Reduce size of ghost.js by moving related Routing code to their own files
