+++ core/server/routes/index.js
@@ -1,0 +1,9 @@
+var api         = require('./api'),
+    admin       = require('./admin'),
+    frontend    = require('./frontend');
+
+module.exports = {
+    api: api,
+    admin: admin,
+    frontend: frontend
+};
